New items, major changes to item code, and more
This is a major update to the code. It has lots of new items and objects
(and liquids!), and makes a lot of changes to the code. Here is the full
list of changes:

- the isEquipped variable is replaced with Inventory.contains(this)
- fixed items "bouncing" when on ground
- platforms now tile properly, with customizable border color and
texture
- a new Timer class that uses delta time (game objects and AI can use
this class to control when different actions are performed)
- separated liquids and platforms from objects
- added lava, bombs and bows (with arrows)
- separated items into different categories (useful for providing custom
functionality for all the different types, like projectiles, explosives,
etc)
- added guns
- started work on a mini nuke launcher
- started work on basic enemy AI

There is more to come, and some of the code still needs work. Stay
tuned!

+++ gui/Overlay.java
@@ -1,8 +1,8 @@
 package gui;
 
-import java.awt.Point;
-
 import item.Item;
+
+import java.awt.Point;
 
 import org.lwjgl.input.Mouse;
 import org.newdawn.slick.Color;
@@ -10,7 +10,6 @@ import org.newdawn.slick.Graphics;
 import org.newdawn.slick.Image;
 
 import player.Inventory;
-import player.Player;
 import enemy.AI;
 import engine.ObjectList;
 
@@ -77,51 +76,51 @@ public class Overlay {
 		if (Inventory.slotThree != null) {
 			g.drawImage(((Item)Inventory.slotThree).defaultTexture, 102, 15, null);
 		}
-		
+
 		drawToolTip(g);
 
 		g.setColor(Color.white);
 
 	}
-	
+
 	public static void drawToolTip(Graphics g) {
-		
+
 		Point mouseLocation = new Point(Mouse.getX(), 600 - Mouse.getY());
-		
+
 		g.setColor(Color.black);
 
 		try {
 			for (int i = 0; i <= ObjectList.items.size(); i++) {
 				if (((Item)ObjectList.items.get(i)).hitbox.contains(mouseLocation)) {
-					g.drawString(((Item) ObjectList.items.get(i)).getClass().toString(), (int)Mouse.getX(), 600 - (int)Mouse.getY());
+					g.drawString(((Item) ObjectList.items.get(i)).getClass().toString(), Mouse.getX(), 600 - Mouse.getY());
 				}
 
 			}
 		} catch (Exception e) {
 
 		}
-		
+
 		try {
 			for (int i = 0; i <= ObjectList.objects.size(); i++) {
 				if (((object.Object)ObjectList.objects.get(i)).hitbox.contains(mouseLocation)) {
-					g.drawString(((object.Object) ObjectList.objects.get(i)).getClass().toString(), (int)Mouse.getX(), 600 - (int)Mouse.getY());
+					g.drawString(((object.Object) ObjectList.objects.get(i)).getClass().toString(), Mouse.getX(), 600 - Mouse.getY());
 				}
 
 			}
 		} catch (Exception e) {
 
 		}
-		
+
 		try {
 			for (int i = 0; i <= ObjectList.enemies.size(); i++) {
 				if (((AI)ObjectList.enemies.get(i)).hitbox.contains(mouseLocation)) {
-					g.drawString(((AI) ObjectList.enemies.get(i)).getClass().toString(), (int)Mouse.getX(), 600 - (int)Mouse.getY());
+					g.drawString(((AI) ObjectList.enemies.get(i)).getClass().toString(), Mouse.getX(), 600 - Mouse.getY());
 				}
 
 			}
 		} catch (Exception e) {
 
 		}
-		
+
 	}
 }
